requirements/base.txt
# Base
pytz==2020.5
python-slugify==4.0.1
Pillow==8.1.0
psycopg2==2.8.6 --no-binary psycopg2

# Django
Django==3.1.5

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.12.2
django-filter==2.4.0
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.7.0

# JWT
pyjwt==2.0.0

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==5.2.0

# Celery
tornado>=6.0.4
redis>=3.5.3
django-redis==4.12.1
celery==4.4.7
flower==0.9.5